requirements.txt
asgiref==0.14.0
autobahn==0.16.0
channels==0.17.2
daphne==0.15.0
defusedxml==0.4.1
Django==1.10.1
django-braces==1.9.0
djangorestframework==3.4.6
oauthlib==1.0.3
PyJWT==1.4.2
python-social-auth==0.2.21
python3-openid==3.0.10
requests==2.11.1
requests-oauthlib==0.6.2
shortuuid==0.4.3
six==1.10.0
Twisted==16.4.0
txaio==2.5.1
zope.interface==4.3.1
